Split up other file.
