+++ cspell.json
@@ -23,6 +23,7 @@
     "clawdrill",
     "clsx",
     "Crewneck",
+    "drinkware",
     "Epik",
     "esque",
     "Famdom",
