+++ src/data/bookmarks.ts
@@ -150,6 +150,19 @@ const data: IItem[] = [
     description:
       "Features the quote \"Forget a man's name and he'll forgive you. Remember it, and he'll defend you forever.\"",
     link: 'https://www.etsy.com/listing/1110349846/morning-star-inspired-bookmark'
+  },
+  {
+    date: '2023-03-14',
+    image: 'https://i.etsystatic.com/16046340/r/il/ca538f/4144213339/il_794xN.4144213339_n2sc.jpg',
+    name: 'Home - Wolf Bookmark',
+    tags: ['other', 'roque'],
+    shop: {
+      name: 'CreativeLeighCrafts',
+      url: 'https://www.etsy.com/shop/CreativeLeighCrafts'
+    },
+    description:
+      'Book mark with "Home isn\'t where you\'re from, it\'s where you find light when all grows dark." on one side and "Hic Est Lupus Motherfucker" on the other.',
+    link: 'https://www.etsy.com/listing/1289914959/home-wolf-bookmark'
   }
 ];
