+++ src/data/candles.ts
@@ -1,0 +1,42 @@
+import { IItem } from '@/types/item';
+
+const data: IItem[] = [
+  {
+    date: '2023-03-27',
+    image: 'https://i.etsystatic.com/17490952/r/il/612087/2775387517/il_794xN.2775387517_h67z.jpg',
+    name: 'Red Rising Inspired Candles',
+    tags: ['other', 'darrow', 'sevro', 'mustang'],
+    shop: {
+      name: 'firenationcreations',
+      url: 'https://www.etsy.com/shop/firenationcreations'
+    },
+    description: 'Choose from Darrow, Mustang, or Sevro candles in 6 oz or 4 oz jars.',
+    link: 'https://www.etsy.com/listing/903972668/red-rising-inspired-candles'
+  },
+  {
+    date: '2023-03-27',
+    image: 'https://i.etsystatic.com/17490952/r/il/90c117/2830626924/il_794xN.2830626924_nybs.jpg',
+    name: 'Song of Persephone Candle',
+    tags: ['other', 'eo'],
+    shop: {
+      name: 'firenationcreations',
+      url: 'https://www.etsy.com/shop/firenationcreations'
+    },
+    description: "Red Rising candle from firenationcreations's Stuck in Space box.",
+    link: 'https://www.etsy.com/listing/952004495/stuck-in-space-sci-fi-box-candles?variation0=1787670255'
+  },
+  {
+    date: '2023-03-29',
+    image: 'https://i.etsystatic.com/18861441/r/il/6e4cca/3412295655/il_794xN.3412295655_leco.jpg',
+    name: 'Red Rising Themed Candles',
+    tags: ['other', 'darrow', 'eo', 'sevro', 'mustang'],
+    shop: {
+      name: 'InTheBooksWeFindUs',
+      url: 'https://www.etsy.com/shop/InTheBooksWeFindUs'
+    },
+    description: 'Red Rising themed candles. Choose from Darrow, Eo, Sevro, or Mustang.',
+    link: 'https://www.etsy.com/listing/765659952/candele-a-tema-red-rising-2-oz'
+  }
+];
+
+export default data;
