+++ src/data/data.ts
@@ -1,5 +1,6 @@
 import bookmarks from './bookmarks';
 import books from './books';
+import candles from './candles';
 import clothing from './clothing';
 import digital from './digital';
 import discs from './discs';
@@ -8,6 +9,7 @@ import hoodies from './hoodies';
 import jewelry from './jewelry';
 import magnets from './magnets';
 import mugs from './mugs';
+import notebooks from './notebooks';
 import other from './other';
 import patches from './patches';
 import phoneCases from './phone-cases';
@@ -15,11 +17,14 @@ import pins from './pins';
 import prints from './prints';
 import stickers from './stickers';
 import sweatshirts from './sweatshirts';
-import waterBottles from './water-bottles';
+
+import drinkware from './drinkware/drinkware';
+import waterBottles from './drinkware/water-bottles';
 
 export const ALL_DATA = [
   ...bookmarks,
   ...books,
+  ...candles,
   ...clothing,
   ...digital,
   ...discs,
@@ -27,6 +32,7 @@ export const ALL_DATA = [
   ...hoodies,
   ...jewelry,
   ...magnets,
+  ...notebooks,
   ...mugs,
   ...other,
   ...patches,
@@ -35,5 +41,5 @@ export const ALL_DATA = [
   ...prints,
   ...stickers,
   ...sweatshirts,
-  ...waterBottles
-];
+  [...drinkware, ...waterBottles]
+].flat();
