+++ src/data/drinkware/drinkware.ts
@@ -1,0 +1,18 @@
+import { IItem } from '@/types/item';
+
+const data: IItem[] = [
+  {
+    date: '2023-03-26',
+    image: 'https://i.etsystatic.com/14027028/r/il/beb6f8/1528102094/il_794xN.1528102094_g267.jpg',
+    name: 'I Would Have Lived in Peace Wine Glass',
+    tags: ['drinkware', 'darrow'],
+    shop: {
+      name: 'EnchantedFandom',
+      url: 'https://www.etsy.com/shop/EnchantedFandom'
+    },
+    description: 'A glittered wine glass with high quality vinyl.',
+    link: 'https://www.etsy.com/listing/620537679/red-rising-by-pierce-brown-glittered'
+  }
+];
+
+export default data;
