+++ src/data/masks.ts
@@ -1,0 +1,34 @@
+import { IItem } from '@/types/item';
+
+const data: IItem[] = [
+  {
+    date: '2023-03-13',
+    image:
+      'https://cdn.shopify.com/s/files/1/0374/1499/4053/products/all-over-print-premium-face-mask-white-front-612516771ce0b_1024x1024.jpg',
+    name: 'Sophocles Face Mask',
+    tags: ['other', 'official'],
+    shop: {
+      name: 'Lit Escalates',
+      url: 'https://www.lit-escalates.com/'
+    },
+    description:
+      "With the Delta Variant raging, we figure it was time to spice up the ol' mask collection. We asked Alexis over at Nerdy Ink to help out and she came through like a boss with this adorable Sophocles illustration.",
+    link: 'https://www.lit-escalates.com/collections/apparel/products/sophocles-face-mask'
+  },
+  {
+    date: '2023-03-13',
+    image:
+      'https://cdn.shopify.com/s/files/1/0374/1499/4053/products/all-over-print-premium-face-mask-black-front-61251a63d6da4_1024x1024.jpg',
+    name: 'Iron Gold Map Face Mask',
+    tags: ['other', 'iron-gold', 'official'],
+    shop: {
+      name: 'Lit Escalates',
+      url: 'https://www.lit-escalates.com/'
+    },
+    description:
+      "With the Delta Variant raging we figured it was time to spice up the ol' mask collection.",
+    link: 'https://www.lit-escalates.com/collections/apparel/products/iron-gold-map-face-mask'
+  }
+];
+
+export default data;
