+++ src/data/notebooks.ts
@@ -1,0 +1,104 @@
+import { IItem } from '@/types/item';
+
+const data: IItem[] = [
+  {
+    date: '2023-03-15',
+    image: 'https://i.etsystatic.com/21279369/r/il/b4db55/3089728740/il_794xN.3089728740_nmlh.jpg',
+    name: 'Ares Riding a Unicorn Spiral Notebook',
+    tags: ['notebook', 'sevro'],
+    shop: {
+      name: 'TheMightyPenBookTees',
+      url: 'https://www.etsy.com/shop/TheMightyPenBookTees'
+    },
+    description: 'A 100 page notebook with a cover featuring Ares riding a unicorn.',
+    link: 'https://www.etsy.com/listing/1007323758/ares-riding-a-unicorn-spiral-notebook'
+  },
+  {
+    date: '2023-03-15',
+    image: 'https://i.etsystatic.com/21279369/r/il/cde421/3089726076/il_794xN.3089726076_apt9.jpg',
+    name: 'Sleepy Sophocles Spiral Notebook',
+    tags: ['notebook'],
+    shop: {
+      name: 'TheMightyPenBookTees',
+      url: 'https://www.etsy.com/shop/TheMightyPenBookTees'
+    },
+    description: 'A 100 page notebook with a cover featuring a sleepy Sophocles.',
+    link: 'https://www.etsy.com/listing/1021299251/sleepy-sophocles-spiral-notebook'
+  },
+  {
+    date: '2023-03-15',
+    image: 'https://i.etsystatic.com/21279369/r/il/05f0e5/3137447025/il_794xN.3137447025_pp8q.jpg',
+    name: 'Live for More Spiral Notebook',
+    tags: ['notebook', 'eo'],
+    shop: {
+      name: 'TheMightyPenBookTees',
+      url: 'https://www.etsy.com/shop/TheMightyPenBookTees'
+    },
+    description: 'A 100 page notebook with a cover featuring Eo and the words "Live for More".',
+    link: 'https://www.etsy.com/listing/1007320876/live-for-more-spiral-notebook'
+  },
+  {
+    date: '2023-03-15',
+    image: 'https://i.etsystatic.com/21279369/r/il/629444/3089719988/il_794xN.3089719988_s7a9.jpg',
+    name: 'Hic Sunt Leones Spiral Notebook',
+    tags: ['notebook'],
+    shop: {
+      name: 'TheMightyPenBookTees',
+      url: 'https://www.etsy.com/shop/TheMightyPenBookTees'
+    },
+    description:
+      'A 100 page notebook with a cover featuring a lion and the words "Hic Sunt Leones".',
+    link: 'https://www.etsy.com/listing/1021297643/hic-sunt-leones-spiral-notebook'
+  },
+  {
+    date: '2023-03-15',
+    image: 'https://i.etsystatic.com/21279369/r/il/7cc179/3089699702/il_794xN.3089699702_rqls.jpg',
+    name: 'Per Aspera Ad Astra Spiral Notebook',
+    tags: ['notebook'],
+    shop: {
+      name: 'TheMightyPenBookTees',
+
+      url: 'https://www.etsy.com/shop/TheMightyPenBookTees'
+    },
+    description: 'A 100 page notebook with a cover with "Per Aspera Ad Astra".',
+    link: 'https://www.etsy.com/listing/1021292115/per-aspera-ad-astra-spiral-notebook'
+  },
+  {
+    date: '2023-03-15',
+    image: 'https://i.etsystatic.com/21279369/r/il/814bad/3137433845/il_794xN.3137433845_4mu5.jpg',
+    name: 'Howler Spiral Notebook',
+    tags: ['notebook'],
+    shop: {
+      name: 'TheMightyPenBookTees',
+      url: 'https://www.etsy.com/shop/TheMightyPenBookTees'
+    },
+    description: 'A 100 page notebook with a wolf on its cover.',
+    link: 'https://www.etsy.com/listing/1021293771/howler-spiral-notebook'
+  },
+  {
+    date: '2023-03-15',
+    image: 'https://i.etsystatic.com/21279369/r/il/545a57/3089710138/il_794xN.3089710138_plky.jpg',
+    name: 'Howler GhostCloak Edition Spiral Notebook',
+    tags: ['notebook'],
+    shop: {
+      name: 'TheMightyPenBookTees',
+      url: 'https://www.etsy.com/shop/TheMightyPenBookTees'
+    },
+    description: 'A 100 page notebook with a wolf on its cover. Special GhostCloak Edition.',
+    link: 'https://www.etsy.com/listing/1007319110/howler-ghostcloak-edition-spiral'
+  },
+  {
+    date: '2023-03-15',
+    image: 'https://i.etsystatic.com/21279369/r/il/1f109e/3137437137/il_794xN.3137437137_qgux.jpg',
+    name: 'S is for Sophocles Spiral Notebook',
+    tags: ['notebook'],
+    shop: {
+      name: 'TheMightyPenBookTees',
+      url: 'https://www.etsy.com/shop/TheMightyPenBookTees'
+    },
+    description: 'A 100 page notebook with a cover featuring Sophocles in the letter S.',
+    link: 'https://www.etsy.com/listing/1021294489/s-is-for-sophocles-spiral-notebook'
+  }
+];
+
+export default data;
